+++ src/components/All/styles.module.css
@@ -1,0 +1,6 @@
+.albumName {
+  margin-right: 1rem;
+}
+.slideTo {
+  margin-left: 1rem;
+}
